Add files via upload
USE CASE -1

The National Weather Service computes the wind-chill index using the following formula:
35.74 + 0.6215*T - 35.75*(V^0·16) + 0.4275*T*(V^0·16)

Where T is the temperature in degrees Fahrenheit, and V is the wind speed in miles per hour.
Write a program that prints a nicely formatted table of wind-chill values. Rows should represent wind speed for 0 to 50 in 5-mph increments, and the columns represent temperatures from -20 to + 60 in 10-degree increments.
Note: The formula only applies for wind speeds in excess of 3 miles per hour.